Make scope build on arm.
